Condition object

Captures what the student passes into pass_if() or fail_if(), figures out what type of object was passed into x, and returns a grader_condition object that will be passed into evaluate_condition().

pass_if and fail_if are wrappers around condition() that sets the proper correct argument. This allows the instructor to specify just the comparison code, x, and message.

condition(x, message, correct)

pass_if(x, message = NULL)

fail_if(x, message = NULL)

Arguments

x
A formula, function, or value, that returns TRUE or FALSE. When comparing objects that are greater than length 1 (e.g., vectors, dataframes, matricies, etc) A boolean vector will be returned if the user uses ==, not a single boolean value. gradethis will run the vector through all(..., na.rm = TRUE) to check for the boolean value. It is advised that the user use identical() instead of == in this case.

message
character string for message returned (usually passed in from pass_if() or fail_if()

correct
logical whether the condition is the correct answer

Value

a grader_condition object that contains the expression x, the message message, whether or not the expression is the correct answer or not, correct, the type of expression (formula, function, or value), type

Functions

condition: A condition object that contains the expression, message, correct status, and condition type

pass_if: a condition that if matched means the student provided result is correct

fail_if: a condition that if matched means the student provided result is incorrect

See also

Examples

# NOT RUN {
gradethis_demo()
# }

condition(~ identical(x = .result, 5), message = "Correct", correct = TRUE)

#> $x
#> ~identical(x = .result, 5)
#> <environment: 0x58231c8>
#> 
#> $message
#> [1] "Correct"
#> 
#> $correct
#> [1] TRUE
#> 
#> $type
#> [1] "formula"
#> 
#> attr(,"class")
#> [1] "grader_condition"

grade_result(
  pass_if(~ identical(.result, 5), "This is a correct message"),
          grader_args = list(),
          learnr_args = list(last_value = 5, envir_prep = new.env())
)

#> $message
#> Super job! This is a correct message 
#> 
#> $correct
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "grader_graded"

grade_result(
  pass_if(function(x) identical(x, 5), "This is a correct message"),
  learnr_args = list(last_value = 5)
)

#> $message
#> Impressive work! This is a correct message 
#> 
#> $correct
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "grader_graded"

grade_result(
  pass_if(5, "This is a correct message"),
          learnr_args = list(last_value = 5)
)

#> $message
#> You should be proud. This is a correct message 
#> 
#> $correct
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "grader_graded"

grade_result(
  fail_if(5, "You were supposed to do this other thing!"),
          pass_if(~ TRUE, "should never reach here"),
          learnr_args = list(last_value = 5)
)

#> $message
#> You were supposed to do this other thing!  But no need to fret, try it again.
#> 
#> $correct
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "grader_graded"