Code Feedback

Generate a message describing the first instance of a code mismatch. This function is built to be used within grade_this() without using arguments. Manual calling of code_feedback() is also encouraged!

code_feedback(
  user_code = get0(".user_code", parent.frame()),
  solution_code = get0(".solution_code", parent.frame()),
  env = get0(".envir_prep", parent.frame(), ifnotfound = parent.frame()),
  ...,
  allow_partial_matching = getOption("gradethis.code.partial_matching", TRUE)
)

maybe_code_feedback(
  user_code = get0(".user_code", parent.frame()),
  solution_code = get0(".solution_code", parent.frame()),
  env = get0(".envir_prep", parent.frame(), ifnotfound = parent.frame()),
  ...,
  allow_partial_matching = getOption("gradethis.code.partial_matching", TRUE),
  default = "",
  space_before = TRUE,
  space_after = FALSE
)

Arguments

user_code
String containing user code. Defaults to retrieving .user_code from the calling environment. (Required)

solution_code
String containing solution code. Defaults to retrieving .solution_code from the calling environment. (Required)

env
Environment used to standardise formals of the user and solution code. Defaults to retrieving .envir_prep from the calling environment. If not found, the parent.frame() will be used

...
Ignored

allow_partial_matching
A logical if FALSE don't allow partial matching

default
Default value to return if no code feedback is found or code feedback can be provided

space_before, space_after
Logical value to determine if a space should be included before (TRUE) or after (FALSE)

Value

If no discrepencies are found, a NULL value is returned. If a code difference is found, a character value describing the difference.

Functions

code_feedback: Determine code feedback

maybe_code_feedback: Return code_feedback() result when possible. Useful when setting default fail() glue messages. For example, if there is no solution, no code feedback will be given.

Code differences

There are many different ways that code can be different, yet still the same. Here is how we detect code differences:

If the single values are different. Ex: log(2) vs log(3)

If the function call is different. Ex: log(2) vs sqrt(2)

Validate the user code can be standardised via rlang::call_standardise(). The env parameter is important for this step as gradethis does not readily know about user defined functions. Ex: read.csv("file.csv") turns into read.csv(file = "file.csv")

If multiple formals are matched. Ex: read.csv(f = "file.csv") has f match to file and fill.

Verify that every named argument in the solution appears in the user code. Ex: If the solution is read.csv("file.csv", header = TRUE), header must exist.

Verify that the user did not supply extra named arguments to .... Ex: mean(x = 1:10, na.rm = TRUE) vs mean(x = 1:10)

Verify that every named argument in the solution matches the value of the corresponding user argument. Ex: read.csv("file.csv", header = TRUE) vs read.csv("file.csv", header = FALSE)

Verify that the remaining aruguments of the user and solution code match in order and value. Ex: mean(1:10, 0.1) vs mean(1:10, 0.2)

Examples

# Values are same
code_feedback("log(2)", "log(2)") # NULL # no differences found

#> NULL

# Functions are different
code_feedback("log(2)", "sqrt(2)")

#> I expected you to call `sqrt()` where you called `log()`.

# Standardise user names
code_feedback("read.csv('file.csv')", "read.csv(file = 'file.csv')") # NULL

#> NULL

# No partial matching
code_feedback("read.csv(f = 'file.csv')", "read.csv(file = 'file.csv')")

#> `read.csv()` accepts more than one argument name that begins with `f`. As a result, R cannot figure out which argument you want to pass `"file.csv"` to. Check how you spelled `f`, or write out the full argument name.

# All named arguments are provided (even if they match default value)
code_feedback("read.csv('file.csv')", "read.csv('file.csv', header = TRUE)")

#> Your call to `read.csv()` should include `"header"` as one of its arguments. You may have misspelled an argument name, or left out an important argument.

# All named argument values match
code_feedback("read.csv('file.csv', header = FALSE)", "read.csv('file.csv', header = TRUE)")

#> In `read.csv("file.csv", header = FALSE)`, I expected `header = TRUE` where you wrote `header = FALSE`.

# No extra arguments are provided
code_feedback("mean(1:10)", "mean(1:10, na.rm = TRUE)")

#> Your call to `mean()` should include `"na.rm"` as one of its arguments. You may have misspelled an argument name, or left out an important argument.

# Unstandardised arguments match in order and value
code_feedback("mean(1:10, 0.1)", "mean(1:10, 0.2)")

#> In `mean(1:10, 0.1)`, I expected `0.2` where you wrote ``:`()`.