Grade result of exercise code

Compares the final result of the student code to known pass_if() and fail_if() condition()s. If the student result exactly matches a known case, returns the matching message value.

grade_result(..., correct = NULL, incorrect = NULL,
  grader_args = list(), learnr_args = list(),
  glue_correct = getOption("gradethis_glue_correct"),
  glue_incorrect = getOption("gradethis_glue_incorrect"))

Arguments

...
pass_if() or fail_if() condition()s to check

correct
A character string to display if the student answer matches a known correct answer.

incorrect
A character string to display if the student answer matches a known incorrect answer.

grader_args
A list of parameters passed to grader functions (provided by grade_learnr()). This contains:

user_quo: Quoted R code submitted by the user. For example rlang::quo(1)

solution_quo: (Optional) Quoted solution R code provided by the *-solution chunk for an exercise.

learnr_args
A list of all parameters passed to grade_learnr() by learnr. See https://rstudio.github.io/learnr/exercises.html#exercise_checking for more details.

glue_correct
A glue string that returns the final correct message displayed. Defaults to getOption("gradethis_glue_correct").

glue_incorrect
A glue string that returns the final incorrect message displayed. Defaults to getOption("gradethis_glue_incorrect").

Value

a graded() object from either pass_if() or fail_if() containing a formatted correct or incorrect message and whether or not a match was found.

See also

Examples

# NOT RUN {
gradethis_demo()
# }

grade_result(
  pass_if(~ identical(.result, 5), "This is a correct message"),
          grader_args = list(),
          learnr_args = list(last_value = 5, envir_prep = new.env())
)

#> $message
#> Bravo! This is a correct message 
#> 
#> $correct
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "grader_graded"

grade_result(
  pass_if(function(x) identical(x, 5), "This is a correct message"),
  learnr_args = list(last_value = 5)
)

#> $message
#> Bravo! This is a correct message 
#> 
#> $correct
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "grader_graded"

grade_result(
  pass_if(5, "This is a correct message"),
          learnr_args = list(last_value = 5)
)

#> $message
#> Superb work! This is a correct message 
#> 
#> $correct
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "grader_graded"

grade_result(
  fail_if(5, "You were supposed to do this other thing!"),
          pass_if(~ TRUE, "should never reach here"),
          learnr_args = list(last_value = 5)
)

#> $message
#> You were supposed to do this other thing!  Give it another try.
#> 
#> $correct
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "grader_graded"