Grade expression

Grade expression

grade_this(
  expr,
  ...,
  fail_code_feedback = getOption("gradethis.fail_code_feedback", TRUE)
)

Arguments

expr
Expression to be evaluated. MUST either signal a grade via pass() or fail() like gradethis functions or throw an error (via testthat or stop()). Errors will be converted to fail() calls with the corresponding error message.

...
ignored

fail_code_feedback
Logical that determines if code feedback should be appended to the default failure message. Can also use maybe_code_feedback() in your custom fail() glue message.

Value

a function whose first parameter should be an environment that contains all necessary information to execute the expression. The evaluation of the expression should return a graded() object.

Available variables

grade_this() allows instructors to create an expression to grade. Within the expression, all learnr tutorial variables variables are available for inspection with a . prepended to the name:

.label: Label for exercise chunk

.solution_code: Code provided within the "-solution" chunk for the exercise

.user_code: R code submitted by the user

.check_code: Code provided within the "-check" (or "-code-check") chunk for the exercise.

.envir_prep: A copy of the R environment before the execution of the chunk.

.envir_result: The R environment after the execution of the chunk.

.evaluate_result: The return value from the evaluate::evaluate function.

`.last_value The last value from evaluating the user's exercise submission. In addition, gradethis has provided some extra variables:

.user, .result: A direct copy of .last_value for friendlier naming

.solution: When accessed, will be the result of evaluating the .solution_code in a child environment of .envir_prep

As the instructor, you are free to use any logic to determine a student's grade as long as a graded() object is signaled. The check code can also contain testthat expectation code. Failed testthat expectations will be turned into fail()ed grades with the corresponding message.

See also

Examples

# These are manual examples, see grading demo for `learnr` tutorial usage

grade_this({
  fail_if_equal(4, "Try adding 1")
  pass_if_equal(5, "You got 5, great!")
  fail()
})(list(
  .result = 4
))

#> <gradethis_graded: [Incorrect] Try adding 1>

grade_this({
  testthat::expect_type(.result, "integer") # will `fail()` if not an integer
  testthat::expect_equal(.result, 5L)        # will `fail()` if not equal to 5
  pass() # returns default message
})(list(
  .result = 5L
))

#> <gradethis_graded: [Correct] Well done! Correct!>

grade_this({
  testthat::expect_true(is.function(.result))
  testthat::expect_equal(.result(1), 3)
  pass()
})(list(
  .result = function(x) {x + 2}
))

#> <gradethis_graded: [Correct] Beautiful! Correct!>

# Remember, only `grade_this(expr)` should be used.
# The followup `list()` and values will be called by `grade_learnr()`
# To learn more about using `grade_this()` with learnr, see:
if (FALSE) gradethis_demo()