Grade student code against a solution

Checks the code expression or the code result against a solution.

grade_this_code(
  correct = getOption("gradethis.code.correct", getOption("gradethis.pass",
    "Correct!")),
  incorrect = getOption("gradethis.code.incorrect", getOption("gradethis.fail",
    "Incorrect")),
  ...,
  allow_partial_matching = getOption("gradethis.code.partial_matching", TRUE)
)

Arguments

correct
A glue-able character string to display if the student answer matches a known correct answer.

incorrect
A glue-able character string to display if the student answer does not match the known correct answer. Use code_feedback() in this string to control the placement of the auto-generated feedback message produced by comparing the student's submission with the solution.

...
Ignored

allow_partial_matching
A logical if FALSE don't allow partial matching

Value

a function whose first parameter should be an environment that contains all necessary information to compare the code. The result of the returned function will be a graded() object.

Details

grade_this_code() compares student code to a solution (i.e. model code) and describes the first way that the student code differs. If the student code exactly matches the solution, grade_this_code() returns a customizable success message (correct). If the student code does not match the solution, a customizable incorrect message (incorrect) can also be provided.

grade_this_code() only inspects for code differences between the student's code and the solution code. The final result of the student code and solution code is ignored. See code_feedback() for implementation details on how code is determined to be different.

You can provide solution code for grade_this_code() to use in two ways:

Pass code as a character string or a quoted expression to the solution argument of grade_code()

Make a "-solution" code chunk for the exercise to be checked in a learnr document. There is no need to supply a solution argument for grade_code() if you call it from the "-check" chunk of the same exercise. Likewise, there is no need to supply a student submitted code argument when you call grade_code() from a learnr document (learnr will provide the code that the student submits when it runs grade_code().

For best results, name all arguments provided in the solution code.

See also

Examples

# These are manual examples, see grading demo for `learnr` tutorial usage

grade_this_code()(list(
  .user_code = "sqrt(log(2))",
  .solution_code = "sqrt(log(1))"
))

#> <gradethis_graded: [Incorrect] I see that you are using pipe operators (e.g. %>%), so I want to let you know that this is how I am interpretting your code before I check it:
#> 
#> ```r
#> sqrt(log(2))
#> ```
#> In `log(2)`, I expected `1` where you wrote `2`. But no need to fret, try it again.>

grade_this_code()(list(
  .user_code = rlang::quo(sqrt(log(2))),
  .solution_code = rlang::quo(sqrt(log(1)))
))

#> Warning: Using `as.character()` on a quosure is deprecated as of rlang 0.3.0.
#> Please use `as_label()` or `as_name()` instead.
#> This warning is displayed once per session.

#> <gradethis_graded: [Incorrect] I see that you are using pipe operators (e.g. %>%), so I want to let you know that this is how I am interpretting your code before I check it:
#> 
#> ```r
#> ~sqrt(log(2))
#> ```
#> In `log(2)`, I expected `1` where you wrote `2`. Try it again. Perseverence is the key to success.>

# Remember, only `grade_this_code(correct, incorrect)` should be used.
# The followup `list()` and values will be called by `grade_learnr()`.
# To learn more about using `grade_this_code()` with learnr, see:
if (FALSE) gradethis_demo()