Graded object for submission value

The return value from graded should be returned by every *-check chunk when used with grade_learnr().

graded(correct, message = NULL)

pass(message = getOption("gradethis.pass", "Correct!"), env = parent.frame())

fail(message = getOption("gradethis.fail", "Incorrect"), env = parent.frame())

pass_if_equal(
  y,
  message = getOption("gradethis.pass", "Correct!"),
  x = get(".result", envir = env),
  env = parent.frame()
)

fail_if_equal(
  y,
  message = getOption("gradethis.fail", "Incorrect"),
  x = get(".result", envir = env),
  env = parent.frame()
)

Arguments

correct
A logical value of whether or not the checked code is correct.

message
A character string of the message to be displayed.

env
environment to evaluate the glue message

y
Required value in which to compare x. Comparison is done using testthat::compare(x, y)

x
First item in the comparison. Typically, this will be the user result

Details

graded() objects are signaled to the calling functions.

grade_result() ignores when grades are created. graded() objects must be returned

grade_code() handles graded() objects internally

grade_this() will stop execution once a pass(), pass_if_equal(), fail(), fail_if_equal(), or grade() is called. To generate programmatic grades, use graded() or if statements around pass*() and fail*()

Functions

graded: Programatic function to produce a graded a result.

pass: Produce a passing grade

fail: Produce a failing grade

pass_if_equal: Produce a passing grade only if testthat::compare() passes on x and y

fail_if_equal: Produce a failing grade only if testthat::compare() passes on x and y