Setup gradethis for use within learnr

Use this function to change the default options suggested by gradethis. This function also describes in detail each of the global options available for customization in the gradethis package. Note that you most likely do not want to change the defaults values for the learnr tutorial options that are prefixed with exercise.. Each of the gradethis-specific arguments sets a global option with the same name, prefixed with gradethis.. For example, pass sets gradethis.pass.

gradethis_setup(
  pass = NULL,
  fail = NULL,
  ...,
  code_correct = NULL,
  code_incorrect = NULL,
  fail_code_feedback = NULL,
  pipe_warning = NULL,
  allow_partial_matching = NULL,
  exercise.checker = gradethis_exercise_checker,
  exercise.timelimit = NULL,
  exercise.error.check.code = NULL
)

Arguments

pass
Default message for pass(). Sets options("gradethis.pass")

fail
Default message for fail(). Sets options("gradethis.fail")

...
Arguments passed on to learnr::tutorial_options

exercise.cap

Caption for exercise chunk (defaults to the engine's icon or the combination of the engine and " code").

exercise.eval

Whether to pre-evaluate the exercise so the reader can see some default output (defaults to FALSE).

exercise.lines

Lines of code for exercise editor (defaults to the number of lines in the code chunk).

exercise.completion

Use code completion in exercise editors.

exercise.diagnostics

Show diagnostics in exercise editors.

exercise.startover

Show "Start Over" button on exercise.

exercise.reveal_solution

Whether to reveal the exercise solution if a solution chunk is provided.

code_correct
Default correct message for grade_this_code(). If unset, grade_this_code() falls back to the value of the gradethis.pass option. Sets options("gradethis.code_correct").

code_incorrect
Default incorrect message for grade_this_code(). If unset grade_this_code() falls back to the value of the gradethis.fail option. Sets options("gradethis.code_incorrect").

fail_code_feedback
Logical TRUE or FALSE to determine whether maybe_code_feedback() should return code feedback, where if FALSE, maybe_code_feedback() will return an empty string. This can be useful in pass() or fail() messages and is used by default in the default fail() message and the default grade_this_code() incorrect message.

pipe_warning
The default message used in pipe_warning(). Sets options("gradethis.pipe_warning").

allow_partial_matching
Logical TRUE or FALSE to determine whether partial matching is allowed in grade_this_code(). Sets options("gradethis.allow_partial_matching").

exercise.checker
Function used to check exercise answers (e.g., gradethis::grade_learnr()).

exercise.timelimit
Number of seconds to limit execution time to (defaults to 30).

exercise.error.check.code
A string containing R code to use for checking code when an exercise evaluation error occurs (e.g., "gradethis::grade_code()").

Value

Invisibly returns the global options as they were prior to setting them with gradethis_setup().

See also

Examples

# Not run in package documentation because this function changes global opts
if (FALSE) {
  old_opts <- gradethis_setup(
    pass = "Great work!", 
    fail = "{random_encouragement()}"
  )
}

# Use getOption() to see the default value
getOption("gradethis.pass")

#> [1] "{random_praise()} Correct!"

getOption("gradethis.code_correct")

#> NULL