Setup gradethis for use within learnr

Call this function inside a learnr document's setup chunk in order to use a suggested behavior when grading exercises.

gradethis_setup(
  exercise.timelimit = 60,
  exercise.checker = grade_learnr,
  exercise.error.check.code = "grade_this_code()",
  ...,
  pass = "{ random_praise() } Correct!",
  fail = "Incorrect.{ maybe_code_feedback() } { random_encouragement() }",
  code_correct = NULL,
  code_incorrect = "{ .message } { random_encouragement() }",
  fail_code_feedback = TRUE,
  allow_partial_matching = NULL
)

Arguments

exercise.timelimit
Number of seconds to limit execution time to (defaults to 30).

exercise.checker
Function used to check exercise answers (e.g., gradethis::grade_learnr()).

exercise.error.check.code
A string containing R code to use for checking code when an exercise evaluation error occurs (e.g., "gradethis::grade_code()").

...
arguments passed to learnr::tutorial_options()

pass
Default message for pass(). Sets options("gradethis.pass")

fail
Default message for fail(). Sets options("gradethis.fail")

code_correct
Sets options("gradethis.code.correct") which read before options("gradethis.pass") for the default correct message in grade_this_code()

code_incorrect
Sets options("gradethis.code.incorrect") which read before options("gradethis.fail") for the default incorrect message in grade_this_code()

fail_code_feedback
Allows for maybe_code_feedback() to return code feedback. (To be paired with a pass() or fail() message.)

allow_partial_matching
Allows for partial matching in grade_this_code(). Sets options("gradethis.code.partial_matching")

See also