praise {gradethis}	R Documentation

Random praise and encouragement

Description

Generate a random praise or encouragement phrase. These functions are designed for use within pass() or fail() messages, or anywhere else that gradethis provides feedback to the student.

Usage

random_praise()

random_encouragement()

give_praise(expr, ..., location = "before", before = NULL, after = NULL)

give_encouragement(expr, ..., location = "after", before = NULL, after = NULL)

Arguments

expr
A graded() grade or helper function, or a grading function — like grade_this() or grade_result() — or a character string. Praise will be added to any passing grades and encouragement will be added to any failing grade. If expr is a character string, then "{random_praise()}" or "{random_encouragement()}" is pasted before or after the string according to location.

...
Ignored.

location
Should the praise or encouragement be added before or after the grade message?

before, after
Text to be added before or after the praise or encouragement phrase.

Value

random_praise() and random_encouragement() each return a length-one string with a praising or encouraging phrase.

give_praise() and give_encouragement() add praise or encouragement phrases to passing and failing grades, respectively.

Functions

random_praise: Random praising phrase

random_encouragement: Random encouraging phrase

give_praise: Add praising message to a passing grade.

give_encouragement: Add encouraging message to a failing grade.

Examples

replicate(5, glue::glue("Random praise: {random_praise()}"))
replicate(5, glue::glue("Random encouragement: {random_encouragement()}"))

# give_praise() adds praise to passing grade messages
give_praise(pass("That's absolutely correct."))

# give_encouragement() encouragement to failing grade messages
give_encouragement(fail("Sorry, but no."))

[Package gradethis version 0.2.5.9000 Index]